Adds a little script that makes anchor links not change address bar.
This is necessary because the inner Aurelia app needs full control of the address bar in order for its routing system to work with links.  Essentially it just makes all clicks on links that start with `#` do a javascript scroll and not an anchor scroll, thereby not changing the address bar.